+++ gismap/lab/labmap.py
@@ -92,8 +92,7 @@ class LabMap(MixInIO):
         for author in tqdm(self._author_iterator(), desc=desc):
             if not all(f(author) for f in self.author_selectors):
                 continue
-            if len(author.sources) == 0:
-                author.auto_sources(dbs=list_of_objects(self.dbs, db_dict(), default=default_dbs))
+            author.auto_sources(dbs=list_of_objects(self.dbs, db_dict(), default=default_dbs))
             if author.sources:
                 self.authors[author.key] = author
             if author.metadata.img is None: